+++ age/src/format/ssh_ed25519.rs
@@ -26,7 +26,7 @@ fn ssh_tag(pubkey: &[u8]) -> [u8; TAG_LEN_BYTES] {
 }
 
 #[derive(Debug)]
-pub(crate) struct RecipientStanza {
+pub struct RecipientStanza {
     pub(crate) tag: [u8; TAG_LEN_BYTES],
     pub(crate) rest: super::x25519::RecipientStanza,
 }