+++ rage/src/bin/rage-mount/main.rs
@@ -199,7 +199,12 @@ fn main() -> Result<(), Error> {
             )?;
 
             decryptor
-                .decrypt_with_callbacks(identities.into_iter(), &UiCallbacks)
+                .decrypt_with_callbacks(
+                    identities
+                        .into_iter()
+                        .map(|i| Box::new(i) as Box<dyn age::Identity>),
+                    &UiCallbacks,
+                )
                 .map_err(|e| e.into())
                 .and_then(|stream| mount_stream(stream, types, mountpoint))
         }